+++ packages/login-link/config/login-link.php
@@ -75,6 +75,10 @@ return [
                 */
             ],
         ],
+        'process' => [
+            'single' => 'Link process',
+            'plural' => 'Link processes',
+        ],
     ],
 
     /*