+++ packages/login-link/src/Resources/LoginLinkProcessResource/Pages/CreatePage.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Moox\LoginLink\Resources\LoginLinkProcessResource\Pages;
+
+use Filament\Resources\Pages\CreateRecord;
+use Moox\LoginLink\Resources\LoginLinkProcessResource;
+
+class CreatePage extends CreateRecord
+{
+    protected static string $resource = LoginLinkProcessResource::class;
+}