+++ packages/login-link/src/Resources/LoginLinkProcessResource/Pages/ListPage.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Moox\LoginLink\Resources\LoginLinkProcessResource\Pages;
+
+use Filament\Actions\CreateAction;
+use Filament\Resources\Pages\ListRecords;
+use Moox\LoginLink\Resources\LoginLinkProcessResource;
+use Override;
+
+class ListPage extends ListRecords
+{
+    protected static string $resource = LoginLinkProcessResource::class;
+
+    #[Override]
+    protected function getHeaderActions(): array
+    {
+        return [
+            CreateAction::make(),
+        ];
+    }
+}